Split stress/olap_workload into several files
