+++ ydb/tests/stress/olap_workload/workload/type/insert_delete.py
@@ -1,0 +1,73 @@
+# -*- coding: utf-8 -*-
+import ydb
+import time
+import random
+import threading
+from enum import Enum
+
+from ydb.tests.stress.common.common import WorkloadBase
+
+class WorkloadInsertDelete(WorkloadBase):
+    def __init__(self, client, prefix, stop):
+        super().__init__(client, prefix, "insert_delete", stop)
+        self.inserted = 0
+        self.current = 0
+        self.table_name = "table"
+        self.lock = threading.Lock()
+
+    def get_stat(self):
+        with self.lock:
+            return f"Inserted: {self.inserted}, Current: {self.current}"
+
+    def _loop(self):
+        table_path = self.get_table_path(self.table_name)
+        self.client.query(
+            f"""
+                CREATE TABLE `{table_path}` (
+                id Int64 NOT NULL,
+                i64Val Int64,
+                PRIMARY KEY(id)
+                )
+                PARTITION BY HASH(id)
+                WITH (
+                    STORE = COLUMN
+                )
+        """,
+            True,
+        )
+        i = 1
+        while not self.is_stop_requested():
+            self.client.query(
+                f"""
+                INSERT INTO `{table_path}` (`id`, `i64Val`)
+                VALUES
+                ({i * 2}, {i * 10}),
+                ({i * 2 + 1}, {i * 10 + 1})
+            """,
+                False,
+            )
+
+            self.client.query(
+                f"""
+                DELETE FROM `{table_path}`
+                WHERE i64Val % 2 == 1
+            """,
+                False,
+            )
+
+            actual = self.client.query(
+                f"""
+                SELECT COUNT(*) as cnt, SUM(i64Val) as vals, SUM(id) as ids FROM `{table_path}`
+            """,
+                False,
+            )[0].rows[0]
+            expected = {"cnt": i, "vals": i * (i + 1) * 5, "ids": i * (i + 1)}
+            if actual != expected:
+                raise Exception(f"Incorrect result: expected:{expected}, actual:{actual}")
+            i += 1
+            with self.lock:
+                self.inserted += 2
+                self.current = actual["cnt"]
+
+    def get_workload_thread_funcs(self):
+        return [self._loop]
