+++ ydb/tests/stress/olap_workload/workload/type/ya.make
@@ -1,0 +1,12 @@
+PY3_LIBRARY()
+
+PY_SRCS(
+    tables_create_drop.py
+    insert_delete.py
+)
+
+PEERDIR(
+    ydb/tests/stress/common
+)
+
+END()
