+++ ydb/tests/stress/olap_workload/workload/ya.make
@@ -9,6 +9,7 @@ PEERDIR(
     ydb/public/sdk/python/enable_v3_new_behavior
     ydb/tests/stress/common
     ydb/tests/datashard/lib
+    ydb/tests/stress/olap_workload/workload/type
 )
 
 END()
